backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
websockets==12.0
opencv-python-headless==4.10.0.84
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
